% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LazyABC.R
\docType{package}
\name{foo}
\alias{foo}
\alias{foo-package}
\title{LazyABC: A package to acompany an OxWaSP report.}
\description{
This package is intended to accompany an OxWaSP report and is not intended
for general distribution.
}

