% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ImportanceABC.R
\name{ImportanceABC}
\alias{ImportanceABC}
\title{Importance Approximate Bayesian Computation}
\usage{
ImportanceABC(data, rimport, dimport, prior, simulator, tolerance = 0.1,
  n = 10000)
}
\arguments{
\item{data}{an observation}

\item{rimport}{a function that samples from the importance distribution}

\item{dimport}{a function taking theta and returning its density from the
importance distribution}

\item{prior}{a function taking theta and returning its density from the prior}

\item{simulator}{a function taking theta performing simulation}

\item{tolerance}{the epsilon tolerance to accept simulations}

\item{n}{number of iterations to run}
}
\value{
n weighted samples of the parameter from the posterior distribution.
}
\description{
Perform importance ABC sampling on user specified distributions.
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}

