% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RejectionABC.R
\name{RejectionABC}
\alias{RejectionABC}
\title{Rejection Approximate Bayesian Computation}
\usage{
RejectionABC(data, prior, simulator, tolerance = 0.1, n = 10000)
}
\arguments{
\item{data}{an observation}

\item{prior}{a function taking theta and returning its density from the prior}

\item{simulator}{a function taking theta performing simulation}

\item{tolerance}{the epsilon tolerance to accept simulations}

\item{n}{number of iterations to run}
}
\value{
n weighted samples of the parameter from the posterior distribution.
}
\description{
Perform rejection ABC sampling on user specified distributions.
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
posterior <-
RejectionABC(data = 2, prior = function(){runif(1, -10, 10)},
             simulator = function(theta){rnorm(1, 2*(theta+2)*theta*(theta-2), 0.1+theta^2)},
             n = 3000, tolerance = 1)
}

